requirements.txt
openai
langchain
streamlit
python-dotenv
PyPDF2
langchain-openai
datetime
langchain-community
Gradio

-e .